% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_matrix_inverse.R
\name{calculate_matrix_inverse}
\alias{calculate_matrix_inverse}
\title{Matrix Inversion}
\usage{
calculate_matrix_inverse(matrix)
}
\arguments{
\item{matrix}{A square matrix.}
}
\value{
The inverse of the input matrix.
}
\description{
This function calculates the inverse of a square matrix using the Gauss-Jordan elimination method.
}
\examples{
matrix <- matrix(c(2, 3, 1, 4), nrow = 2)
inverse_matrix <- calculate_matrix_inverse(matrix)
# Returns:  0.8 -0.2
#          -0.6  0.4

}
