% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{Correlation}
\alias{Correlation}
\title{Computes the correlation between two vectors}
\usage{
Correlation(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A list with the correlation coefficient and the description of the correlation.
}
\description{
This function calculates the correlation between two numeric vectors
using the \code{cor} function in R. It also returns a list that contains both
the correlation coefficient and the description of the correlation.
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(2, 4, 6, 8, 10)
Correlation(x, y)
Returns: List(correlation = 1, description = "Positive correlation")
}
