% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_regression.R
\name{polynomial_regression}
\alias{polynomial_regression}
\title{Polynomial Regression}
\usage{
polynomial_regression(x, y, degree)
}
\arguments{
\item{x}{A numeric vector of predictor variable values.}

\item{y}{A numeric vector of response variable values.}

\item{degree}{The degree of the polynomial model to fit.}
}
\value{
A list containing the polynomial coefficients and the predicted values.
}
\description{
This function performs polynomial regression on a given dataset, fitting a polynomial model of specified degree.
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(2, 4, 6, 8, 10)
result <- polynomial_regression(x, y, degree = 2)
result$coefficients
# Expected Output: (Intercept)   predictors1   predictors2
#                  0.000000e+00  2.000000e+00 -1.718126e-16
result$predicted_values
# Expected Output: 1  2  3  4  5
#                  2  4  6  8 10

}
