% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fibonacci.R
\name{calculate_fibonacci}
\alias{calculate_fibonacci}
\title{Calculate Fibonacci Sequence}
\usage{
calculate_fibonacci(limit)
}
\arguments{
\item{limit}{An integer specifying the maximum value in the Fibonacci sequence.}
}
\value{
A numeric vector containing the Fibonacci sequence.
}
\description{
This function generates the Fibonacci sequence up to a specified limit.
}
\examples{
calculate_fibonacci(10)
# Returns: 0, 1, 1, 2, 3, 5, 8

}
